add buttons instead of using view controllers
