--- passBackText copy/passBackText.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Workspace
-   version = "1.0">
-   <FileRef
-      location = "self:passBackText.xcodeproj">
-   </FileRef>
-</Workspace>
